+++ app/Http/Controllers/Admin/BranchController.php
@@ -19,8 +19,12 @@ class BranchController extends Controller
      */
     public function index()
     {
-        $branches = Branch::latest()->get();
-        return view('admin.branch.index', compact('branches'));
+        if (auth()->user()->can('view_branch')) {
+                
+                $branches = Branch::latest()->get();
+                return view('admin.branch.index', compact('branches'));
+            }
+        abort(403);
     }
 
     /**
@@ -30,8 +34,12 @@ class BranchController extends Controller
      */
     public function create()
     {
-        $areas = Area::all();
-        return view('admin.branch.create', compact('areas'));
+        if (auth()->user()->can('create_branch')) {
+                
+                $areas = Area::all();
+                return view('admin.branch.create', compact('areas'));
+            }
+        abort(403); 
     }
 
     /**
@@ -42,18 +50,22 @@ class BranchController extends Controller
      */
     public function store(BranchStoreRequest $request)
     {
-        /* create branch */
-        $branch = Branch::create([
-            'name' => $request->branch,
-            'slug' => str_slug($request->branch),
-            'area_id' => $request->area,
-        ]);
-        /* cheack and showing toastr message */
-        if($branch){
-            Toastr::success('Branch Successfully Added', 'Success');
-            return redirect()->route('admin.branch.index');
-        }
-        abort(404);
+        if (auth()->user()->can('create_branch')) {
+                        /* create branch */
+                $branch = Branch::create([
+                    'name' => $request->branch,
+                    'slug' => str_slug($request->branch),
+                    'area_id' => $request->area,
+                ]);
+                /* cheack and showing toastr message */
+                if($branch){
+                    Toastr::success('Branch Successfully Added', 'Success');
+                    return redirect()->route('admin.branch.index');
+                }
+                abort(404);
+                
+            }
+        abort(403);
     }
 
     /**
@@ -75,8 +87,12 @@ class BranchController extends Controller
      */
     public function edit(Branch $branch)
     {
-        $areas = Area::all();
-        return view('admin.branch.edit', compact('branch', 'areas'));
+       if (auth()->user()->can('edit_branch')) {
+               
+                $areas = Area::all();
+                return view('admin.branch.edit', compact('branch', 'areas'));
+           }
+       abort(403);
     }
 
     /**
@@ -88,20 +104,24 @@ class BranchController extends Controller
      */
     public function update(BranchUpdateRequest $request, Branch $branch)
     {
-         /* update branch */
-         $resultBranch = $branch->update([
-            
-            'area_id'  =>   $request->area,
-            'name'     =>   $request->branch,
-            'slug'     =>   str_slug($request->branch),
-        ]);
+         if (auth()->user()->can('edit_branch')) {
+                 
+                         /* update branch */
+                 $resultBranch = $branch->update([
+                    
+                    'area_id'  =>   $request->area,
+                    'name'     =>   $request->branch,
+                    'slug'     =>   str_slug($request->branch),
+                ]);
 
-        /* cheack and showing toastr message */
-        if($resultBranch){
-            Toastr::success('Branch Successfully Updated', 'Success');
-            return redirect()->route('admin.branch.index');
-        }
-        abort(404);
+                /* cheack and showing toastr message */
+                if($resultBranch){
+                    Toastr::success('Branch Successfully Updated', 'Success');
+                    return redirect()->route('admin.branch.index');
+                }
+                abort(404);
+             }
+         abort(403);
     }
 
     /**
@@ -112,8 +132,12 @@ class BranchController extends Controller
      */
     public function destroy(Branch $branch)
     {
-        $branch->delete();
-        Toastr::success('Branch Successfully Deleted', 'Success');
-        return redirect()->route('admin.branch.index');
+        if (auth()->user()->can('delete_branch')) {
+                
+                $branch->delete();
+                Toastr::success('Branch Successfully Deleted', 'Success');
+                return redirect()->route('admin.branch.index');
+            }
+        abort(403);
     }
 }
