+++ app/Http/Controllers/Admin/CategoryController.php
@@ -19,9 +19,13 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::latest()->get();
+        if (auth()->user()->can('view_category')) {
+                
+                $categories = Category::latest()->get();
 
-        return view('admin.category.index', compact('categories'));
+                 return view('admin.category.index', compact('categories'));
+            }
+        abort(403);
          
     }
 
@@ -32,7 +36,11 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('admin.category.create');
+        if (auth()->user()->can('create_category')) {
+                
+                return view('admin.category.create');
+            }
+        abort(403);
     }
 
     /**
@@ -43,17 +51,21 @@ class CategoryController extends Controller
      */
     public function store(CategoryStoreRequest $request)
     {
-        /* category created */
-        $category = Category::create([
-            'name' => $request->category,
-            'slug' => str_slug($request->category),
-        ]);
-        
-        if($category) {
-            Toastr::success('Category Successfully Inserted', 'Success');
-            return redirect()->route('admin.category.index');
-        }
-        abort(404);
+        if (auth()->user()->can('create_category')) {
+                            /* category created */
+                $category = Category::create([
+                    'name' => $request->category,
+                    'slug' => str_slug($request->category),
+                ]);
+                
+                if($category) {
+                    Toastr::success('Category Successfully Inserted', 'Success');
+                    return redirect()->route('admin.category.index');
+                }
+                abort(404);
+                
+            }
+        abort(403);
        
     }
 
@@ -76,7 +88,11 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
-        return view('admin.category.edit', compact('category'));
+        if (auth()->user()->can('edit_category')) {
+                
+                return view('admin.category.edit', compact('category'));
+            }
+        abort(403);
     }
 
     /**
@@ -88,18 +104,22 @@ class CategoryController extends Controller
      */
     public function update(CategoryUpdateRequest $request, Category $category)
     {
-        /* update category */
-        $resultCategory = $category->update([
-            'name'  => $request->category,
-            'slug'  => str_slug($request->category),
-        ]);
+       if (auth()->user()->can('edit_category')) {
+               
+                        /* update category */
+                $resultCategory = $category->update([
+                    'name'  => $request->category,
+                    'slug'  => str_slug($request->category),
+                ]);
 
-        /* check and toastr message */
-        if($resultCategory) {
-            Toastr::success('Category Successfully Updated', 'Success');
-            return redirect()->route('admin.category.index');
-        }
-        abort(404);
+                /* check and toastr message */
+                if($resultCategory) {
+                    Toastr::success('Category Successfully Updated', 'Success');
+                    return redirect()->route('admin.category.index');
+                }
+                abort(404);
+           }
+       abort(403);
         
     }
 
@@ -111,8 +131,12 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
-        Toastr::success('Category Successfully Deleted', 'Success');
-        return redirect()->route('admin.category.index');
+       if (auth()->user()->can('delete_category')) {
+               
+                $category->delete();
+                Toastr::success('Category Successfully Deleted', 'Success');
+                return redirect()->route('admin.category.index');
+           }
+       abort(403);
     }
 }
